+++ bundler/CMakeLists.txt
@@ -6,6 +6,8 @@ set(CMAKE_CXX_STANDARD 17)
 set(SOURCE_FILES
         "src/OutputStream.h"
         "src/OutputStream.cpp"
+        "src/ModuleFile.h"
+        "src/ModuleFile.cpp"
         "src/ModuleResolver.h"
         "src/ModuleResolver.cpp"
         "src/UniqueNameGenerator.h"
